task #refactor controller
